feat: add InMemoryContextFake to seed data in memory database

+++ src/Aiko.Domain/Models/EquipmentPositionHistory.cs
@@ -10,9 +10,9 @@ namespace Aiko.Domain.Models
         public Guid EquipmentId { get; set; }
         [Column("eqph_dt_date", TypeName = "datetime")]
         public DateTime Date { get; set; }
-        [Column("eqph_nm_lat", TypeName = "int")]
-        public int Lat { get; set; }
-        [Column("eqph_nm_lon", TypeName = "int")]
-        public int Lon { get; set; }
+        [Column("eqph_nm_lat", TypeName = "double")]
+        public double Lat { get; set; }
+        [Column("eqph_nm_lon", TypeName = "double")]
+        public double Lon { get; set; }
     }
 }
